Cassandra assets rebuild (#134)
* create new IcoBundle relations

* add mew Ico relations and tables

* add Api annotations

* fix bugs and generate migrations diff

* right diff magrations

* create relations and fix

* migration

* fix tests

* fix tests

* add calculate ETH ico price

* remove extra migration

* scrutinizer fix

* scrutinizer fix

* processing of data from API IcoBench for writing to the database

* create actual migrations diff

* delete dead variable

* scrutinizer fix

* Update .scrutinizer.yml

* fix relation depth and FinanceNormalizer refactor

* try scrutinizer fix

* exclude Ico entity

* create new IcoBundle relations

* add mew Ico relations and tables

* add Api annotations

* fix bugs and generate migrations diff

* right diff magrations

* create relations and fix

* migration

* fix tests

* fix tests

* add calculate ETH ico price

* remove extra migration

* scrutinizer fix

* scrutinizer fix

* processing of data from API IcoBench for writing to the database

* create actual migrations diff

* delete dead variable

* scrutinizer fix

* Update .scrutinizer.yml

* fix relation depth and FinanceNormalizer refactor

* try scrutinizer fix

* exclude Ico entity

* resolve merge conflicts

* resolve conflicts one more

* add Api annotations

* fix bugs and generate migrations diff

* right diff magrations

* create relations and fix

* add calculate ETH ico price

* remove extra migration

* scrutinizer fix

* processing of data from API IcoBench for writing to the database

* fix after rebase

* fix bug

* delete relations with Post entity

* one migration instead two

* sort market cap

* refactor Market cap entity

* checked can be merged

* test

* add check for time

* remove check for time

* chenge procfile

* back to procfile

* fix typo service

* start work

* continue work

* chenge Cassandra tables and logic

* start add history watcher

* add watcher for coins history data

* fix history data watcher

* change procfile for first sart commands

* fix tests

* fix cassandra connection in history for prod

* fix cassandra client in history watcher

* back procfile

* add check for dublicate history data

* flush redis data

* change cassandra migrations for prod

* fix some in redis flushall

* after redis clean

* chamge histoty data array set

* fix dump

* redis flushall

* restart all commands

* fix history data without async request

* fix response

* fix response status code

* fix response status code one more

* back procfile again

+++ src/Kami/AssetBundle/Model/TradableToken.php
@@ -230,9 +230,9 @@ class TradableToken
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function getVolume(): int
+    public function getVolume(): float
     {
         return $this->volume;
     }
